+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - fix: remove unnecessary translations from the connecting-payment-form-without-api page [DV-4160]
+- feat: add Advanced Troubleshooting section and “Re-reading a Block from Blockchain” page (all locales) [DV-4240]
 
 ## [0.0.20]